Add signing functions.
The intolerable slowness of the Ed25519 implementation really shows
here.  I don't _really_ understand value of signed but unencrypted data
with the ability to use anonymous reicpients in the encryption stage.

Nevertheless, here it is for completeness.  And maybe it will inspire me
to profile and improve the Ed25519 code.